% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chunktop-package.R
\docType{package}
\name{chunktop-package}
\alias{chunktop}
\alias{chunktop-package}
\title{chunktop: Read R Markdown Chunk Options and Reinsert Them}
\description{
Read the options from an R Markdown chunk and prepare them into a data.frame. Present this to colleagues for editing, then read the file back in and reinsert the changes back to the respective R Markdown chunk options. Based on 'parsermd' functions.
}
\author{
\strong{Maintainer}: Matt Dray \email{mwdray@gmail.com}

}
\keyword{internal}
