% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chunktop.R
\name{chunktop_to_df}
\alias{chunktop_to_df}
\title{Convert a List of R Markdown Chunk Options to a Data Frame}
\usage{
chunktop_to_df(chunktop_list)
}
\arguments{
\item{chunktop_list}{Nested list. One element per chunk. One sub-element per
option.}
}
\value{
A data.frame with three columns: 'chunk_name' (character) of the name
of the chunk containing the option of interest, 'opt_name' (character) for
the name of the option of interest and 'opt_value' (character, numeric or
logical) for the value stored by the option represented by that row.
}
\description{
Having extracted specific chunk options from an R Markdown file with
\link{get_chunktop}, convert the list output to a data.frame.
}
\examples{
# Read example R Markdown file
path <- system.file("extdata/demo1.Rmd", package = "chunktop")

# Extract 'fig.cap' chunk from each chunk into a list, convert to data.frame
get_chunktop(path, "fig.cap") |> chunktop_to_df()

}
