% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chunktop.R
\name{df_to_chunktop}
\alias{df_to_chunktop}
\title{Convert a Data Frame of R markdown Chunk Options to a List}
\usage{
df_to_chunktop(chunktop_df)
}
\arguments{
\item{chunktop_df}{Data.frame. Contains chunk information for R Markdown
chunks to be replaced. Columns are 'chunk_name' (character) of the name
of the chunk containing the option of interest, 'opt_name' (character)
for the name of the option of interest and 'opt_value' (character,
numeric or logical) for the value stored by the option represented by
that row.}
}
\value{
A list. One element per chunk, with sub-lists named for each chunk
option.
}
\description{
Convert a Data Frame of R markdown Chunk Options to a List
}
\examples{
# Read example R Markdown file
path <- system.file("extdata/demo1.Rmd", package = "chunktop")

# Create temporary CSV file to hold output
csv_file <- tempfile(fileext = ".csv")

# Extract 'fig.cap' chunk option into a list, convert to data.frame, write to CSV
get_chunktop(path, "fig.cap") |>
 chunktop_to_df() |>
 write.csv(csv.file, row.names = FALSE)

# Read CSV back in and re-convert into list
read.csv(csv_file) |> df_to_chunktop(chunktop_df)

}
