% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chunktop.R
\name{get_chunktop}
\alias{get_chunktop}
\title{Extract Chunk Options}
\usage{
get_chunktop(rmd_file, opts)
}
\arguments{
\item{rmd_file}{Character. Path to an R Markdown (.Rmd) file to be read.}

\item{opts}{Character. A vector of named options to read from R Markdown
chunks.}
}
\value{
A list. One element per chunk, with sub-lists named for each chunk
option named in 'opts', if present.
}
\description{
Parse an R Markdown file and read selected chunk options of interest to a
list, where each list element is named after the chunk.
}
\details{
Each chunk in the target R Markdown file must have a unique chunk
name.
}
\examples{
# Read example R Markdown file
path <- system.file("extdata/demo1.Rmd", package = "chunktop")

# Extract 'fig.cap' chunk from each chunk into a list
get_chunktop(path, "fig.cap")

}
